requirements/development.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/development.txt requirements.in/development.txt
#
attrs==19.3.0             # via pytest
backcall==0.1.0           # via ipython
click==7.0                # via pip-tools
coverage==5.0.3           # via pytest-cov
decorator==4.4.2          # via ipython, traitlets
django-cors-headers==3.2.1  # via -r requirements.in/base.txt
django-dynamic-preferences==1.8.1  # via -r requirements.in/base.txt
django-extensions==2.2.8  # via -r requirements.in/base.txt
django==2.2.10            # via -r requirements.in/base.txt, django-cors-headers, django-dynamic-preferences, djangorestframework, easy-thumbnails
djangorestframework==3.11.0  # via -r requirements.in/base.txt
easy-thumbnails==2.7      # via -r requirements.in/base.txt
entrypoints==0.3          # via flake8
faker==0.9.1              # via mixer
flake8==3.7.9             # via -r requirements.in/development.txt
ipython-genutils==0.2.0   # via traitlets
ipython==7.13.0           # via -r requirements.in/development.txt
jedi==0.16.0              # via ipython
mccabe==0.6.1             # via flake8
mixer==6.1.3              # via -r requirements.in/development.txt
more-itertools==8.2.0     # via pytest
packaging==20.1           # via pytest
parso==0.6.2              # via jedi
persisting-theory==0.2.1  # via django-dynamic-preferences
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==6.2.2             # via -r requirements.in/base.txt, easy-thumbnails
pip-tools==4.5.1          # via -r requirements.in/development.txt
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.3     # via ipython
psycopg2==2.8.4           # via -r requirements.in/base.txt
ptyprocess==0.6.0         # via pexpect
py==1.8.1                 # via pytest
pycodestyle==2.5.0        # via flake8
pyflakes==2.1.1           # via flake8
pygments==2.5.2           # via ipython
pyparsing==2.4.6          # via packaging
pytest-cov==2.8.1         # via -r requirements.in/development.txt
pytest-django==3.8.0      # via -r requirements.in/development.txt
pytest-mock==2.0.0        # via -r requirements.in/development.txt
pytest==5.3.5             # via -r requirements.in/development.txt, pytest-cov, pytest-django, pytest-mock
python-dateutil==2.8.1    # via faker
python-decouple==3.3      # via -r requirements.in/base.txt
pytz==2019.3              # via django
six==1.14.0               # via django-dynamic-preferences, django-extensions, faker, packaging, pip-tools, python-dateutil, traitlets
sqlparse==0.3.1           # via django
text-unidecode==1.2       # via faker
traitlets==4.3.3          # via ipython
wcwidth==0.1.8            # via prompt-toolkit, pytest

# The following packages are considered to be unsafe in a requirements file:
# setuptools
